chore: Split up init_shell into multiple files
